requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
plotly>=5.0.0
dash>=2.0.0

# NVIDIA libraries
pynvml>=11.4.1
nvidia-ml-py3>=7.352.0

# Web framework
flask>=2.0.0
fastapi>=0.70.0
uvicorn>=0.15.0

# Data processing
scipy>=1.7.0
scikit-learn>=1.0.0

# Utilities
tqdm>=4.62.0
click>=8.0.0
pyyaml>=6.0
toml>=0.10.2

# Development dependencies
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.9.0
flake8>=3.9.0
mypy>=0.910

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0

# Jupyter support
jupyter>=1.0.0
jupyterlab>=3.2.0
ipykernel>=6.4.0

# Optional: CUDA Python (if available)
# cupy-cuda11x>=9.0.0  # Uncomment if CUDA 11.x is available
# numba>=0.54.0        # Uncomment for GPU-accelerated NumPy operations

# Optional: Machine learning for advanced features
# tensorflow>=2.7.0    # Uncomment for ML-based optimization
# torch>=1.10.0        # Uncomment for PyTorch integration





